fix(backend-core): Throw on missing JWT key
