+++ packages/backend-core/src/index.ts
@@ -2,7 +2,8 @@ export * from './Base';
 export * from './api/ClerkBackendAPI';
 export * from './api/resources';
 export { createGetToken, createSignedOutState } from './util/createGetToken';
+export { AuthStatus } from './types';
 export type { ClerkFetcher } from './api/utils/RestClient';
 export type { Session } from './api/resources/Session';
 export type { Nullable } from './util/nullable';
-export type { JWTPayload } from './util/types';
+export type { JWTPayload } from './types';
